// web/src/api/generated/schemas/lotMedia.ts
/**
 * Generated by orval v7.20.0 🍺
 * Do not edit manually.
 * Rush Hour Real Estate Platform API
 * MVP платформы недвижимости - публичные эндпоинты
 * OpenAPI spec version: 1.0.0
 */
import type { MediaItem } from './mediaItem'

export interface LotMedia {
  photos?: MediaItem[]
  floorPlanImages?: MediaItem[]
  cover?: MediaItem
}